third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-change-node-while-editing.html

<html>
<head>
<script src="../../../inspector/inspector-test.js"></script>
<script src="../../../inspector/elements-test.js"></script>
<script>

function test() {
  ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step1);

  var treeElement;
  var section;

  function step1() {
    ElementsTestRunner.dumpSelectedElementStyles(true, true, true);
    treeElement = ElementsTestRunner.getElementStylePropertyTreeItem('color');

    treeElement.startEditing();
    treeElement.nameElement.textContent = 'background';

    ElementsTestRunner.selectNodeAndWaitForStyles('other', step2);
  }

  function step2() {
    ElementsTestRunner.dumpSelectedElementStyles(true, true, true);
    ElementsTestRunner.selectNodeAndWaitForStyles('inspected', step3);
  }

  function step3() {
    ElementsTestRunner.dumpSelectedElementStyles(true, true, true);
    TestRunner.completeTest();
  }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that changing selected node while editing style does update styles sidebar.
</p>

<div id="inspected" style="color: red">Text</div>
<div id="other" style="color: blue"></div>

</body>
</html>
