third_party/WebKit/LayoutTests/http/tests/devtools/elements/styles-3/styles-change-node-while-editing-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that changing selected node while editing style does update styles sidebar.

[expanded] 
element.style { ()
    color: red;

[expanded] 
element.style { ()
    color: blue;

[expanded] 
element.style { ()
    background: red;
